requirements.txt
torch
torchvision
pandas
numpy
matplotlib
scikit-learn
Pillow
streamlit